+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Currently no unreleased changes
 
+## [0.18.1] - 2021-08-10
+### Fixed
+- Fix parking position disabling when parking position is gone
+- Fix force parameter for updates
+
 ## [0.18.0] - 2021-08-09
 ### Added
 - Possibility to fetch chargers without updating chargersList
@@ -204,7 +209,8 @@ Minor fix in observer interface
 ## [0.1.0] - 2021-05-26
 Initial release
 
-[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.18.0...HEAD
+[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.18.1...HEAD
+[0.18.1]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.18.1
 [0.18.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.18.0
 [0.17.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.17.0
 [0.16.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.16.0